# docs/swagger.yaml
basePath: /v1
definitions:
...
  entity.CreateBusinessRequestForSwagger:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
...
  entity.UpdateBusinessRequestForSwagger:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
...
  status_http.Response:
    properties:
      custom_message: {}
      data: {}
      description:
        type: string
      status:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Using a translation service as an example
  title: Go Clean Template API
  version: "1.0"
paths:
...
  /business/create:
    post:
      consumes:
      - application/json
      description: Create a new business with the provided details
      parameters:
      - description: Business details
        in: body
        name: business
        required: true
        schema:
          $ref: '#/definitions/entity.CreateBusinessRequestForSwagger'
      produces:
      - application/json
      responses:
        "201":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Create a new business
      tags:
      - BUSINESS
  /business/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a business by ID
      parameters:
      - description: Business ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Delete a business
      tags:
      - BUSINESS
  /business/get/{id}:
    get:
      consumes:
      - application/json
      description: Get business details by its ID
      parameters:
      - description: Business ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Business
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Get a business by ID
      tags:
      - BUSINESS
  /business/list:
    get:
      consumes:
      - application/json
      description: Get all businesses with pagination
      parameters:
      - default: 10
        description: Limit per page
        in: query
        name: limit
        type: integer
      - default: 0
        description: Page number
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/status_http.Response'
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Get all businesses
      tags:
      - BUSINESS
  /business/update/{id}:
    put:
      consumes:
      - application/json
      description: Update business details
      parameters:
      - description: Business ID
        in: path
        name: id
        required: true
        type: string
      - description: Business details
        in: body
        name: business
        required: true
        schema:
          $ref: '#/definitions/entity.UpdateBusinessRequestForSwagger'
      produces:
      - application/json
      responses:
        "200":
          description: Business updated
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Update a business
      tags:
      - BUSINESS
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"